+++ src/main/scala/io/quartz/QuartzH2Server.scala
@@ -398,7 +398,7 @@ class QuartzH2Server(
   def run0(e: ExecutorService, R: HttpRoute, maxThreadNum: Int, maxStreams: Int, keepAliveMs: Int): IO[ExitCode] = {
     for {
       addr <- IO(new InetSocketAddress(HOST, PORT))
-      _ <- Logger[IO].info("HTTP/2 TLS Service: QuartzH2 (async - Java NIO)")
+      _ <- Logger[IO].info("HTTP/2 TLS Service: QuartzH2 async mode (netio)")
       _ <- Logger[IO].info(s"Concurrency level(max threads): $maxThreadNum, max streams per conection: $maxStreams")
       _ <- Logger[IO].info(s"h2 idle timeout: $keepAliveMs Ms")
       _ <- Logger[IO].info(
@@ -445,7 +445,7 @@ class QuartzH2Server(
   def run1(R: HttpRoute, maxThreadNum: Int, maxStreams: Int, keepAliveMs: Int): IO[ExitCode] = {
     for {
       addr <- IO(new InetSocketAddress(HOST, PORT))
-      _ <- Logger[IO].info("HTTP/2 TLS Service: QuartzH2 ( sync - Java Socket )")
+      _ <- Logger[IO].info("HTTP/2 TLS Service: QuartzH2 sync mode (sockets)")
       _ <- Logger[IO].info(s"Concurrency level(max threads): $maxThreadNum, max streams per conection: $maxStreams")
       _ <- Logger[IO].info(s"h2 idle timeout: $keepAliveMs Ms")
       _ <- Logger[IO].info(s"Listens: ${addr.getHostString()}:${addr.getPort().toString()}")
@@ -488,7 +488,7 @@ class QuartzH2Server(
   def run3(e: ExecutorService, R: HttpRoute, maxThreadNum: Int, maxStreams: Int, keepAliveMs: Int): IO[ExitCode] = {
     for {
       addr <- IO(new InetSocketAddress(HOST, PORT))
-      _ <- Logger[IO].info("HTTP/2 h2c service: QuartzH2 (async - Java NIO)")
+      _ <- Logger[IO].info("HTTP/2 h2c service: QuartzH2 async mode (netio)")
       _ <- Logger[IO].info(s"Concurrency level(max threads): $maxThreadNum, max streams per conection: $maxStreams")
       _ <- Logger[IO].info(s"h2c idle timeout: $keepAliveMs Ms")
       _ <- Logger[IO].info(
